Using int32_t instead of long in simplexnoise

+++ src/simplexnoise.cpp
@@ -484,7 +484,7 @@ float dot( const int* g, const float x, const float y, const float z, const floa
 
 #define PI 3.14159265
 
-int simplexnoise(long seed, float* map, int width, int height, float roughness)
+int simplexnoise(int32_t seed, float* map, int width, int height, float roughness)
 {
     float ka = 256/seed;
     float kb = seed*567%256;
